+++ include/file_io/bounding_box_by_timestamp_io.h
@@ -18,10 +18,10 @@ const uint64_t kDefaultCameraId = std::numeric_limits<uint64_t>::max();
 struct BoundingBoxWithTimestampAndId {
   uint64_t ellipsoid_idx;
 
-  uint64_t min_pixel_x;
-  uint64_t min_pixel_y;
-  uint64_t max_pixel_x;
-  uint64_t max_pixel_y;
+  double min_pixel_x;
+  double min_pixel_y;
+  double max_pixel_x;
+  double max_pixel_y;
 
   /**
    * Semantic class of the detected bounding box
@@ -41,10 +41,10 @@ struct BoundingBoxWithTimestampAndId {
 };
 
 struct BoundingBoxWithTimestamp {
-  uint64_t min_pixel_x;
-  uint64_t min_pixel_y;
-  uint64_t max_pixel_x;
-  uint64_t max_pixel_y;
+  double min_pixel_x;
+  double min_pixel_y;
+  double max_pixel_x;
+  double max_pixel_y;
 
   /**
    * Semantic class of the detected bounding box
